Proxyを別のファイルに分割。
